Rename java inference helpers to more explanatory names, and start scaladoc'ing the usage.
